exhibit states for veto

+++ app/Http/Controllers/ExhibitController.php
@@ -14,8 +14,16 @@ class ExhibitController extends Controller
      */
     public function index()
     {
-        $exhibits = Exhibit::withCount('animals')->get();
-        return view('exhibits.index', compact('exhibits'));
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if ($user->isVeterinary()) {
+            $exhibits = Exhibit::withCount('animals')->get();
+            return view('vet.exhibits.index', compact('exhibits'));
+        } elseif ($user->isAdmin()) {
+            $exhibits = Exhibit::withCount('animals')->get();
+            return view('exhibits.index', compact('exhibits'));
+        }
+        abort(403);
     }
 
     /**
@@ -23,11 +31,21 @@ class ExhibitController extends Controller
      */
     public function create()
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin()) {
+            abort(403);
+        }
         return view('exhibits.create');
     }
 
     public function store(Request $request)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin()) {
+            abort(403);
+        }
         $validated = $request->validate([
             'slug' => 'nullable|string|max:255',
             'name' => 'required|string|max:255',
@@ -55,6 +73,11 @@ class ExhibitController extends Controller
      */
     public function edit(Exhibit $exhibit)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin()) {
+            abort(403);
+        }
         return view('exhibits.edit', compact('exhibit'));
     }
 
@@ -64,6 +87,11 @@ class ExhibitController extends Controller
     // public function update(UpdateExhibitRequest $request, Exhibit $exhibit)
     public function update(Request $request, Exhibit $exhibit)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin() && !$user->isVeterinary()) {
+            abort(403);
+        }
         $validated = $request->validate([
             'slug' => 'nullable|string|max:255',
             'name' => 'required|string|max:255',
@@ -82,8 +110,42 @@ class ExhibitController extends Controller
      */
     public function destroy(Exhibit $exhibit)
     {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isAdmin() && !$user->isVeterinary()) {
+            abort(403);
+        }
         $exhibit->delete();
 
         return redirect()->route('exhibits.index')->with('success', 'Exhibit deleted successfully.');
+    }
+
+    public function state(Exhibit $exhibit)
+    {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isVeterinary()) {
+            abort(403);
+        }
+        return view('vet.exhibits.state', compact('exhibit'));
+    }
+
+    public function stateUpdate(Request $request, Exhibit $exhibit)
+    {
+        /** @var \App\Models\User */
+        $user = auth()->user();
+        if (!$user->isVeterinary()) {
+            abort(403);
+        }
+        $validated = $request->validate([
+            'state_at' => 'nullable|date',
+            'state' => 'nullable|string',
+        ]);
+
+        $exhibit->state_at = $validated['state_at'];
+        $exhibit->state = $validated['state'];
+        $exhibit->save();
+
+        return redirect()->route('exhibits.index')->with('success', 'Exhibit state updated successfully.');
     }
 }
